chore: reorganize integration tests
